fix: Move enum classes into a separate file
